Move InputUID into coreComponents directory
