+++ admin/src/components/index.js
@@ -2,7 +2,6 @@ export { default as CreateModal } from './CreateModal';
 export { default as EditMenuItem } from './EditMenuItem';
 export { default as FormLayout } from './FormLayout';
 export { default as Initializer } from './Initializer';
-export { default as InputUID } from './InputUID';
 export { default as Layout } from './Layout';
 export { default as MenuItemsManager } from './MenuItemsManager';
 export { default as MenuManagerProvider } from './MenuManagerProvider';
@@ -12,3 +11,6 @@ export { default as Section } from './Section';
 export { default as Toolbar } from './Toolbar';
 export { default as TreeMenu } from './TreeMenu';
 export { default as TreeMenuItem } from './TreeMenuItem';
+
+// The following components were cloned from Strapi's core files with minor mods.
+export { default as InputUID } from '../coreComponents/InputUID';
