+++ admin/src/coreComponents/InputUID/endActionStyle.js
