+++ admin/src/coreComponents/InputUID/regex.js
